Conflicts:
	External/Plugins/CodeRefactor/Commands/Move.cs
	External/Plugins/CodeRefactor/PluginMain.cs
	PluginCore/PluginCore/Helpers/FileHelper.cs

+++ PluginCore/PluginCore/Controls/CompletionList.cs
@@ -843,9 +843,9 @@ namespace PluginCore.Controls
         /// <summary>
         /// 
         /// </summary> 
-		static public int GetHandle()
+		static public IntPtr GetHandle()
 		{
-			return (int)completionList.Handle;
+			return completionList.Handle;
 		}
 
         /// <summary>
